Move peer metadata type around
