+++ lib/collection/Cargo.toml
@@ -63,7 +63,7 @@ chrono = { version = "~0.4", features = ["serde"] }
 schemars = { workspace = true }
 tar = "0.4.40"
 fs_extra = "1.3.0"
-semver = "1.0.22"
+semver = { workspace = true }
 tempfile = "3.10.1"
 sha2 = "0.10.8"
 bytes = "1.5.0"
