+++ lib/storage/Cargo.toml
@@ -32,7 +32,6 @@ parking_lot = { workspace = true }
 tar = "0.4.40"
 chrono = { version = "~0.4", features = ["serde"] }
 validator = { version = "0.16", features = ["derive"] }
-semver = { workspace = true }
 
 # Consensus related
 atomicwrites = { version = "0.4.3" }
